Clean up of the code related to constructors from Json in several classes. Minor updates.
