+++ sacco/Code/NetworkInfo.cs
@@ -11,6 +11,8 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using Newtonsoft.Json;
+
 
 namespace commercio.sacco.lib
 {
@@ -25,12 +27,14 @@ namespace commercio.sacco.lib
         // Optional fields
         public String name { get; private set; }                // Human readable chain name
         public String iconUrl { get; private set; }             // Chain icon url
-        public String defaultTokenDenom { get; private set; }   
+        public String defaultTokenDenom { get; private set; }
 
         #endregion
 
         #region Constructors
 
+        // I need to mark this constructor in order to avoid confusion in deserialization
+        [JsonConstructor]
         public NetworkInfo(String bech32Hrp, String lcdUrl, String name = "", String iconUrl = "", String defaultTokenDenom = "")
         {
             Trace.Assert(bech32Hrp != null);
@@ -45,22 +49,23 @@ namespace commercio.sacco.lib
             */
         }
 
-        // Static constructor from json - I am not sure this is the best way to do it, I'd prefer a standard constructor from a dictionary - however, this is near to Dart approach
-        public static NetworkInfo fromJson(Dictionary<String, Object> json)
+        // I abandoned static constructor from json - I am not sure this is the best way to do it, I'd prefer a standard constructor from a dictionary - athough, this was near to Dart approach
+        // Std cosntructor from a dictionary - should work!!!
+        public NetworkInfo (Dictionary<String, Object> json)
         {
             Object outValue;
-            String wkBech32Hrp = null, wkLcdUrl = null, wkName = null, wkIconUrl = null, wkDefaultTokenDenom = null;
+            // String wkBech32Hrp = null, wkLcdUrl = null, wkName = null, wkIconUrl = null, wkDefaultTokenDenom = null;
             if (json.TryGetValue("bech32Hrp", out outValue))
-                wkBech32Hrp = outValue as String;
+                this.bech32Hrp = outValue as String;
             if (json.TryGetValue("lcd_url", out outValue))
-                wkLcdUrl = outValue as String;
+                this.lcdUrl = outValue as String;
             if (json.TryGetValue("name", out outValue))
-                wkName = outValue as String;
+                this.name = outValue as String;
             if (json.TryGetValue("icon_url", out outValue))
-                wkIconUrl = outValue as String;
+                this.iconUrl = outValue as String;
             if (json.TryGetValue("default_token_denom", out outValue))
-                wkDefaultTokenDenom = outValue as String;
-            return new NetworkInfo(wkBech32Hrp, wkLcdUrl, wkName, wkIconUrl, wkDefaultTokenDenom);
+                this.defaultTokenDenom = outValue as String;
+            // return new NetworkInfo(wkBech32Hrp, wkLcdUrl, wkName, wkIconUrl, wkDefaultTokenDenom);
         }
