+++ sacco/sacco.csproj
@@ -19,7 +19,6 @@
 
   <ItemGroup>
     <PackageReference Include="BouncyCastle.NetCore" Version="1.8.5" />
-    <PackageReference Include="CompareNETObjects" Version="4.65.0" />
     <PackageReference Include="NBitcoin" Version="5.0.13" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
   </ItemGroup>
